requirements.txt
pybit
torch
pandas
numpy
ccxt
python-dotenv
requests
websocket-client
scikit-learn
tenacity
